vendor payment and ledger

+++ code/protected/migrations/m170329_062017_VendorAccountDetails.php
@@ -6,7 +6,7 @@ class m170329_062017_VendorAccountDetails extends CDbMigration
 	{
         $transaction = Yii::app()->db->beginTransaction();
         try{
-            $sql = 'alter table cb_dev_groots.vendors add column account_holder_name varchar(255) default null, add column bank_account_no int(18) default null, 
+            $sql = 'alter table cb_dev_groots.vendors add column account_holder_name varchar(255) default null, add column bank_account_no BIGINT(18) default null, 
                       add column account_type enum ("Saving", "Current") default "Saving", add column bank_name VARCHAR (300) default null , 
                       add column branch_name VARCHAR (255) default null, add column isfc_code varchar (15)';
             $this->execute($sql);
